refactor code into 3 files
