+++ utils.py
@@ -1,0 +1,92 @@
+import spacy
+import numpy as np
+import matplotlib.pyplot as plt
+from nltk.corpus import wordnet
+from collections import OrderedDict
+
+nlp = spacy.load('en_core_web_sm')
+
+nouns = {x.name().split('.', 1)[0] for x in wordnet.all_synsets('n')}
+verbs = {x.name().split('.', 1)[0] for x in wordnet.all_synsets('v')}
+
+
+def get_synonyms(word, number= 3):
+    synonyms = []
+    for syn in wordnet.synsets(word): 
+        for l in syn.lemmas(): 
+            synonyms.append(l.name().lower().replace("_", " "))
+    synonyms = list(OrderedDict.fromkeys(synonyms))
+    return synonyms[:number]
+
+
+def tokenize(doc):
+    """
+    Returns a list of strings containing each token in `sentence`.
+    """
+    return [token.text for token in nlp.tokenizer(doc)]
+
+
+def preprocess(doc):
+    """
+    Returns tokens excluding stop words.
+    """
+    return " ".join([token.lemma_ for token in nlp(doc) if not token.is_stop])
+
+
+def plot_results(results):
+    """
+    Plots the results from clf.
+    """
+    indices = np.arange(len(results))
+
+    results = [[x[i] for x in results] for i in range(4)]
+
+    clf_names, score, training_time, test_time = results
+    training_time = np.array(training_time) / np.max(training_time)
+    test_time = np.array(test_time) / np.max(test_time)
+
+    plt.figure(figsize=(12, 8))
+    plt.title("Score")
+    plt.barh(indices, score, .2, label="score", color='navy')
+    plt.barh(indices + .3, training_time, .2, label="training time",
+                color='c')
+    plt.barh(indices + .6, test_time, .2, label="test time", color='darkorange')
+    plt.yticks(())
+    plt.legend(loc='best')
+    plt.subplots_adjust(left=.25)
+    plt.subplots_adjust(top=.95)
+    plt.subplots_adjust(bottom=.05)
+
+    for i, c in zip(indices, clf_names):
+        plt.text(-.3, i, c)
+
+    plt.show()
+
+
+def find_ngrams(input_list, n):
+    return zip(*[input_list[i:] for i in range(n)])
+
+
+def semhash_tokenizer(text):
+    tokens = text.split(" ")
+    final_tokens = []
+    for unhashed_token in tokens:
+        hashed_token = "#{}#".format(unhashed_token)
+        final_tokens += [''.join(gram)
+                            for gram in list(find_ngrams(list(hashed_token), 3))]
+    return final_tokens
+
+
+def semhash_corpus(corpus):
+    new_corpus = []
+    for sentence in corpus:
+        sentence = preprocess(sentence)
+        tokens = semhash_tokenizer(sentence)
+        new_corpus.append(" ".join(map(str,tokens)))
+    return new_corpus
+
+
+def trim(s):
+    """Trim string to fit on terminal (assuming 80-column display)"""
+    return s if len(s) <= 80 else s[:77] + "..."
+
